+++ CHANGELOG.md
@@ -7,9 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Add experimental feature to detect resource detectors in auto instrumentation
+  ([#3181](https://github.com/open-telemetry/opentelemetry-python/pull/3181))
+- Fix exporting of ExponentialBucketHistogramAggregation from opentelemetry.sdk.metrics.view
+  ([#3240](https://github.com/open-telemetry/opentelemetry-python/pull/3240))
+- Fix headers types mismatch for OTLP Exporters
+  ([#3226](https://github.com/open-telemetry/opentelemetry-python/pull/3226))
+- Fix suppress instrumentation for log batch processor
+  ([#3223](https://github.com/open-telemetry/opentelemetry-python/pull/3223))
+- Add speced out environment variables and arguments for BatchLogRecordProcessor
+  ([#3237](https://github.com/open-telemetry/opentelemetry-python/pull/3237))
+
+## Version 1.17.0/0.38b0 (2023-03-22)
+
 - Implement LowMemory temporality
   ([#3223](https://github.com/open-telemetry/opentelemetry-python/pull/3223))
-- PeriodicExportingMetricReader will continue if collection times out 
+- PeriodicExportingMetricReader will continue if collection times out
   ([#3100](https://github.com/open-telemetry/opentelemetry-python/pull/3100))
 - Fix formatting of ConsoleMetricExporter.
   ([#3197](https://github.com/open-telemetry/opentelemetry-python/pull/3197))